Track original referrer
In order to grant certain benefits based on a visitors source we need to
keep track of their origin. This is achieved by storing the original
referrer in a cookie shared by the tapglue.com domain so it can be read
by the dashboard. Which in turn puts the original referrer as metadata
in the Organization under `original_referrer` and passes the last
referrer as field in the emitted segment event for the member creation.

Fixes #380