Let auth section be responsive
In order to provide a seamless experience for users on all device types
we ensure adapting style of the auth section (login, signup). In order
to do so we take the grid capabilities of reflex[0] and adhere to its
standards completely. The auth section is tested on all major device
sizes. It is advisable to adapt this method for the rest of the
Dashboard and maybe all other browser based components with shared
styles.

[0] http://leejordan.github.io/reflex/docs/