Introduce copy to of App tokens
With the help of the Clipboard module we allow now for copying of app tokens on
click on the token button. Visual feedback is realised by changing the content
of the button to signal that the copy succeeded and after a delay of 2 seconds
it shows the actual token again.