Account for broken value in EC
We observed occasionally broken values for ratelimites in ElasticCache
where the stored valued would be -1 with a missing TTL. This led to
blocked customers as the API would report zero remaining requests left
for the ratelimit bucket. This change resets the ratelimit in case of a
returned -1. This is not ideal, but will ensure uninterrupted customer
intraction. We do not assume 100% accurrancy for rate limits.