Always create key for ratelimit bucket
So far we used the NX extension on SET which wouldn't create the bucket
if it already exists. This was a fair assumption until we discovered the
behaviour of a decrement on a key that doesn't exist. It leavs the
bucket without a TTL and a value of -1. To make this not affect the rate
limiting we starting of by decrementing the bucket and request the TTL
information afterwards. If this turns out to leave the bucket without an
expiration time we simply create it with the intended quota of the
passed limitee.