Add metrics collection
With the help of the newly created monitoring AMI including Prometheus,
Pushgateway, Alertmanager and Grafana. We're going to bootstrap two
dedicated instances for metrics collection across our infrastructure.
For now this includes service level metrics on port 9000 (the intaker
exposes it) and host level monitoring on port 9100 provided by the
node-exporter. We need to extend the installation base of node-exporter
by updating our AMI. This is out-of-scope of this change and should be
part of the formalised creation of our AMIs.

Part of #486.