Simplify firewall rules
In order to have more flexibility in our inter service communication this
change aims to reduce the amount of security groups. Firstly this is
necessary as there is a limit of 5 per node. More importantly it should
improve separation of concerns and draw a clearer picture of our
infrastructure and should result in an average of 2 to 3 groups per
host. When defining a new node in our VPC it will always have the basic
group `private` applied and additionally the separation in service and
platform, if necessary a third role can be used to express pairing with
a specific component. It's unclear if the third role is necessary or if
we can generalise all concerns into platform offerings.

While it usually is preferable to pair the declaration close to the
usage, it's not the ideal in this scenario as the pattern is absolutely
symmetrical. Hence the groups are defined in one file. Being as follows:

* gateway - Capability to translate public facing protocol and transport
* loadbalancer - Termination and load-shedding of protocol and transport
* private - Baseline for all nodes in our private network
* platform - Components  (owned or provided) used by/for services
* service - Owned components when composed satisfying public interface