Implment objects in first iteration
This is an attempt to provide a generic objects feature. An object is a
user owned entity with a list of attachments. They can target existing
objects or foreign objects (not known to Tapglue). Attachments are
currently limited to text and URL with a maximum amount of 5 per
object. Symmetrically to Events, Objects can have a custom type and have
a visibility which determins what audience can see the object.

The change brings the definition of an object service as well as
concrete implementations against an in-memory and postgres store. Where
all business logic is encapsulated in the controller layer, which as of
now depends on other domain entities. Those are frontend by so called
StrangleServices which are used to provide logging and instrumentation.

A service can be considered production ready when it logs, exposes
metrics and in the future is traced.

All the new object functionality is stitched into the current code by
creating a subrouter of the current main router and define the endpoints
against the `0.4` prefixed subrouter. While that isolated the code path,
it also necessitated the demand for duplicated functionality in the form
of HTTP protocol handling. All concerns around that are express in
gatweay/http.

It strives for this logical separation:

router - translation VERB and PATH to endpoint
middleware - metadata extraction, protocol constraints
handler - paylaod translation into domain
controller - business constraints
service - entity constraints, persistency

list of additions:
* http middlewares
* object handler
* object controller
* object service
* app strangle service
* connection strangle service
* user strangle service