Introduce id_string in the responses
In order to overcome issues with clients that don't support our lengthy IDs
we introduce the _string version for them which should cover all the fields
that we have to represent ids but as string.

The change is visible only to the clients of the API, internally nothing is
changed.