Use json un/marshaler interface for intermediate type
Instead of having to rely of the use of helper functions we equip the
intermediate type used to accept and return Object representations in
HTTP with the expected interface implementations in order for the json
package to work.

This pattern should be preferred when dealing with encoding/decoding in
handler code.