Implement Posts API
Based on objects we are modelling the Tapglue owned entities Posts and
their companion comments. It aims to provide the backbone of
user-generated content (UGC) and will be extended with additional
functionality like Likes and Shares. Major driver was to have a simple
and very opionionated stance on a common use-case: the typical post in a
social network with its major interactions.

* introduce owned field for events and objects
* add Posts CRUD handler
* add PostsController
* add Comments CRUD
* add CommentsController