Harden posts boundary
Instead of passing Objects between controller and handler layer we now
have an intermediate type Post which is used to further clarify the
separation. This will also be the place to encode additional information
in the future (e.g. if the post is liked by the current user).

Question remains if introducing types in the controller layer is the
right way forward or if a Post should be in the realm of the objects
service. A change later down the road should not impose too much
overhead.