Fix connection counts for deleted users
Up until know we'd only retrieve the connections to present the counts.
If a user in the social graph was deleted this wasn't taken into account
as we don't delete connection on user deletion. This change addresses
this issue by first retrieving the user ids for the connection types and
then counting the users which are not deleted of those lists.