Ignore errors in users convenience method
The errors from this convenience method used to bubble up and prevent
the feed construction which is undesirable when the helper is only used
to compose the accompaning user map. Originally the issue obseved was a
deleted user tried to be read into the map. A proper fix for this is to
filter events for existence of the resources they point/refer to.

Fixes #662