Clean feed representation
iWith increased usage of our Feed feature we observe unintended
representations. Those include events which reference entities (users,
objects, connections) which don't exist anymore or duplicated events
for interactions like follow-unfollow in short succession.

Ideally we represent a clean list of entities from a multitude of
sources (social graph, global, second degree social, potentially
interest graph). The challenge presents itself in that we are operating
on isolated buckets of documents without referential integrity. While
this is a delibrate design decision we need to account for it.

As a side-effect we now transport all entities referenced.