Add external interactions
Wherein we describe how to offer the core interactions we offer for
Posts, Comments and Likes (in the future: Votes, Shares, etc.). The UGC
package evolves around the idea of a fully user created corpus of
entities, which ignores a variety of use-cases where customers already
own a large set of entities which easily mapped to Posts. We assume that
the reasons are non-negotiable, while they still want the core
interactions.

We now provide a new set of routes which at their roots share /externals
and offer the same functionality as /posts. Instead of expecting a post
id we allow a set of alphanumeric characters to be passed which is used
to associate the entity created with it for later reference and list
retrieval and counting.