Add Analytics to dashboard
Customers want to know how their apps are performing, As a first step to
make that visible to them is an analytics view in the Dashboard. This
first iteration adds four dimensions per application for the following
entities: connections, events, objects, users.

The dimensions are visualised by day bukets stacked the time span.