Rework service instrumentation and latency
Changes the way we track service metrics by moving from a set of metrics
per service we share the metrics across all services and express the
important dimensions like service name in the labels we track.

Currently it's uncertain if you can track sub-second latency with go-kit
TimeHistogram as it seems to rely on handling absolute numbers, hene we
can't use it to track our latencies which aim to use prometheus
suggested way of tracking latencies in Histograms with percentiles
expressed in relation to a second. A discussion has been started to
address this in go-kit.

* Use shared servie metrics
* Fix latency tracking
* Reduce cardinality explosion when tracking user-agents