Implement comment/like counts on Posts
In our system we store streams of entities which have relations of some
form, think comments on a post. A big part in UI/UX is to display number
of objects especially in the context of interactive elements where user
engagement is desired numebrs can serve as social currency and elemental
backbone of the network.

To support those use-cases we want to start injecting counts for a
limited set of pre-defined items. We're starting off with Posts and the
amount of Comments and Likes as it is one of the most requested addition
to the UGC package. A logical next step would to expose similar numbers
on the User.

Technically we went for the naive way of counting the entries directly
with the knowledge that it could become a problem. To address that the
following steps should help
* analyse and optimise count queries
* transparently mix a read-through cache
* dedicated count service (highly dependent on event based architecture)

Changes:
* Add Count method to event Service (mem, postgres implementation)
* Add Count method to object Service (mem, postgres implementation)
* Enrich posts with comment & like counts
* Expose counts for posts in API