Introduce automated reporting
We want to track the success of our customers and for that we need to
understand how their apps are performing. This change introduces an
automated reporter which outputs key metrics for all apps running in
production to our important communciation channels (Slack for now). And
data storage in Google Spreadsheet and S3 is planned.