Rotate service logs to prevent disk starvation
Deploys started to fail because the disks of our ecs agents filled up
as /var/log/message* wasn't rotated properly and used up multiple
gigbytes. This change introduces an aggressive rotation policy with
compression included. As we store our logs with loggly anyway there is
little use, except for short term forensics, for the logs on the hosts
themselves.