Extend sessions with device scope
Up until now sessions weren't handled in the context of devices/origins
and until recently weren't reused when existed. With the move towards
automated notifications for end user devices we need a more complete
image of the session origin to be able to reason about the validity of a
device with its token over the session lifecycle.

* add device_id to session
* remove session table creation
* extract and pass DeviceID