Add language to devices
In order to be able to send messages tailored to the language the user
uses on their device we accept and store a language field which is
expected to be BCP 47 compliant. While from this change on mandatory
there is no change of behaviour based on it yet.

Part of #683.