Avoid recursive pull of graph
When computing the neighbours for a feed request we pulled in all
adjacent users, instead of filtering for friends and followings only. As
we also compute connection changes for neighbours that would mean that
users with high follower count would be recomputed everytime a feed
request for one of their follower was executed.