Introduce connection query limit
While observing that feed computation can take in the order of multiple
seconds for users with a huge follower count it gets apparent that the
current way we compile list of IDs and enrich it with users is not
efficient. As a short-term improvement until we have proper pagination
we limit the follower numbers to 200 for the injection of `tg_follow`
events.

* introduce `Limit` query option for connections
* limit followers count to 200 in feed response