React to endpoint changes
In order to react to device changes properly we have to incorporate
Amazons feedback loop provided via SNS notifications which then are
hooked up to an SQS queue for inspection.

When we recognise a delivery failure for a device we disable the device
in question which can be reverted with a subsequent update from the
client.