Rewrite App code paths
As one of the last areas in the 0.4 API section we want to get feature
and code parity for all endpoints. In order to do this for Apps this
change-set introduces an App service, endpoint handlers and control
logic for all app CRUD operations and additionally all App handling in
other crucial paths (e.g. app extraction from authentication headers).

* introduce App service
* implement postgres app service
* rewrite App endpoint handlers
* implement App controller logic
* remove all Object related endpoints