Add post_id to event payload
To reach consistency with our entity abstractions we are adding the
`post_id` field to event payloads which is the same as `tg_object_id`.
This aims to improve the ease of interaction with the API and should
improve natural correlation between referenced posts and posts delivered
in auxiliary maps.