Move to pganalyze Go collector
In order to benefit from better introspection and potential early access
to features in their beta program we move from the python based
collector scheduled by data pipelines in eu-west to the rewrite done in
Go. This gives us the opportunity to run it as normal ECS service and
reduces the operational complexity.