Implement post interactions restrictions
In order to give limited control over interactions on posts we want to
introduce a set of flags which indicate if restrictions are present for
a given Post. To not extend the API surface and make potential
integration and migration more complicated we want limit this change to
an extension of the Post data model itself. We also intrdouce a flag for
report for external use and future-provability. Setting these should
only be allowed for backend tokens.

```
{
  "restrictions": {
    "comment": true,
    "like": false,
    "report": true
  }
}
```