Add approval process state information to comment
As part of community management we understand user generated content to
be subject to an approval/review process. We envision this to be in form
of reports created by users and reviewed by moderators. To enable
similar use-cases for customers in the meantime we allow state
information to be stored for comments.