Implement cached event counts
With the increased usage of the API we understand that the majority of
interactions with the database are counts of secondary entities e.g.
lieks on a Post. The result of this computation is not time sensitive
and the change of rate doesn't warrant a repeated computation if we can
relax the expectations around those counts.

This change-set addresses the issue by caching the results of the count
queries in redis for the current default of 5 minutes (300 seconds). By
means of implement a service middleware for the events service which
intercepts the Count method. The main short-coming with this particular
implementation is that visibility is lacking as in service timings for
the communication with Redis and logging of actual commands initiated.
Ideally we want the same visibility as with a Service wrapped with
logging and instrumentation, which requires a revisit of the current
setup and must be evaluated carefully. We assume this is good enough for
now as the actual impact is not entirely understood until put in
production, so we treat it as experiment which can be extended or
dropped entirely given the results.

* implement cache service middleware
* wire up cache middleware in gateway-http