Fix uniqueness constrain on devices
Up until now we used device_id and user_id to check for uniqueness,
which is not reliable when installation of an app from mixed sources
takes place as the device_id might change and the token stays the same.
This will result in multiple registered devices and lead to notification
duplicates.